+++ src/main/java/com/web/billim/product/domain/ProductInterest.java
@@ -18,7 +18,7 @@ public class ProductInterest {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "interest_id")
-    private int interestId;
+    private long interestId;
 
     @JoinColumn(name = "product_id", referencedColumnName = "product_id")
     @ManyToOne
