+++ src/main/java/com/web/billim/product/dto/response/MyProductSalesResponse.java
@@ -13,10 +13,10 @@ import java.util.stream.Collectors;
 @Builder
 public class MyProductSalesResponse {
 
-	private int productId;
+	private long productId;
 	private String productName;
 	private String detail;
-	private int price;
+	private long price;
 	private List<String> imageUrls;
 	private List<TradeMethod> tradeMethods;
