+++ src/main/java/com/web/billim/product/dto/response/ProductListResponse.java
@@ -13,11 +13,11 @@ import java.util.stream.Collectors;
 @Builder
 public class ProductListResponse {
 
-    private int productId;
+    private long productId;
     private String productName;
     private ProductCategory productCategory;
     private String detail;
-    private int price;
+    private long price;
     private List<String> imageUrls;
     private double starRating;
